+++ is_valid/is_if.py
@@ -1,3 +1,5 @@
+import warnings
+
 from .base import Predicate
 from .to_pred import to_pred
 
@@ -13,7 +15,7 @@ class is_if(Predicate):
     """
 
     def __init__(
-        self, condition, if_predicate, else_predicate=None, else_valid=False
+        self, condition, if_predicate, else_predicate=None, else_valid=True
     ):
         self._cond = to_pred(condition)
         self._if = to_pred(if_predicate)
@@ -21,6 +23,15 @@ class is_if(Predicate):
             None if else_predicate is None else to_pred(else_predicate)
         )
         self._else_valid = else_valid
+
+        if not self._else_valid:
+            warnings.warn(
+                (
+                    'Using is_if with else_valid=False is deprecated, use '
+                    'is_pre instead.'
+                ),
+                DeprecationWarning, stacklevel=2,
+            )
 
     def _evaluate(self, data, explain, context):
         res = self._cond(data, explain and self._else is None, context)