+++ is_valid/is_iterable_of.py
@@ -16,19 +16,32 @@ class is_iterable_of(Predicate):
         self._predicate = to_pred(predicate)
 
     def _evaluate_explain(self, data, context):
-        reasons, errors = {}, {}
+        reasons = {}
+        errors = {}
+        new_data = []
+
         for i, value in enumerate(data):
             explanation = self._predicate.explain(value, context)
-            (reasons if explanation else errors)[i] = explanation
-        return Explanation(
-            True, 'iterable_of',
-            'all elements are valid according to the predicate',
-            reasons,
-        ) if not errors else Explanation(
-            False, 'not_iterable_of',
-            'not all elements are valid according to the predicate',
-            errors,
-        )
+            if explanation:
+                reasons[i] = explanation
+            else:
+                errors[i] = explanation
+            new_data.append(explanation.data)
+
+        if errors:
+            explanation = Explanation(
+                False, 'not_iterable_of',
+                'not all elements are valid according to the predicate',
+                errors,
+            )
+        else:
+            explanation = Explanation(
+                True, 'iterable_of',
+                'all elements are valid according to the predicate',
+                reasons,
+            )
+        explanation.data = new_data
+        return explanation
 
     def _evaluate_no_explain(self, data, context):
         return all(