+++ is_valid/is_tuple_where.py
@@ -12,3 +12,9 @@ class is_tuple_where(is_iterable_where):
     """
 
     prerequisites = [is_tuple]
+
+    def _evaluate(self, data, explain, context):
+        res = super()._evaluate(data, explain, context)
+        if explain:
+            res.data = tuple(res.data)
+        return res