+++ tests/test_condition_predicates.py
@@ -4,7 +4,7 @@ from hypothesis import given
 import hypothesis.strategies as hs
 
 from is_valid import is_any, is_all, is_one, is_if, is_cond, is_something,\
-    is_nothing, is_eq
+    is_nothing, is_eq, is_pre
 
 
 class TestConditionPredicates(TestCase):
@@ -46,18 +46,29 @@ class TestConditionPredicates(TestCase):
         with self.subTest('explain=True == explain=False'):
             self.assertEqual(pred(1), pred.explain(1).valid)
         with self.subTest('pred correct'):
-            self.assertEqual(pred(1), subpred(1) if cond(1) else False)
+            self.assertEqual(pred(1), subpred(1) if cond(1) else True)
 
     @given(
         hs.sampled_from([is_something, is_nothing]),
         hs.sampled_from([is_something, is_nothing])
     )
     def test_is_if_else_valid(self, cond, subpred):
-        pred = is_if(cond, subpred, else_valid=True)
+        pred = is_if(cond, subpred, else_valid=False)
         with self.subTest('explain=True == explain=False'):
             self.assertEqual(pred(1), pred.explain(1).valid)
         with self.subTest('pred correct'):
-            self.assertEqual(pred(1), subpred(1) if cond(1) else True)
+            self.assertEqual(pred(1), subpred(1) if cond(1) else False)
+
+    @given(
+        hs.sampled_from([is_something, is_nothing]),
+        hs.sampled_from([is_something, is_nothing])
+    )
+    def test_is_pre(self, cond, subpred):
+        pred = is_pre(cond, subpred)
+        with self.subTest('explain=True == explain=False'):
+            self.assertEqual(pred(1), pred.explain(1).valid)
+        with self.subTest('pred correct'):
+            self.assertEqual(pred(1), subpred(1) if cond(1) else False)
 
     @given(
         hs.sampled_from([is_something, is_nothing]),